Why:
The final core model is Subscription, the interaction between User and
Trip.

When a user texts a trip ID, it will create a subscription, and
the user will get notifications for that subscription for the duration
of the subscription.

We will need each trip's subscribers so we know who to notify when a
trip gets updated.

This change addresses the need by:

* Adding a subscription model.

* Updating User so we can access each user's subscriptions and trips.

* Updating Trip so we can access each trip's subscriptions
and subscribers.